Add workaround for non-vanishing correlation potential
There seem to be noisy values in the correlation potential
for cases where the electron density of one spin channel is
zero (e.g. H atom). Fluctuating values in the correlation
potential have confused the mixer and made SCF unstable. In
the affected cases, the correlation potential of the empty
spin channel is now manually set to zero, which drastically
improves SCF convergence. This only affects the solution of
the empty spin channel, i.e. not, for example, the total
energy or eigenvalues of the occupied states.